refactor: consolidate IndexedDB utility functions and remove deprecated code
